Use WebApplicationBuilder
Move to using WebApplicationBuilder using a daily build of .NET 6 preview 6.
